Add files via upload
RTO Program - In this Program firstly we have to enter the number like ( MH05CH2565 ) to know the exact information saved in the rto database the program have to fetch the information from API KEY and it will give the total no details present their.      NOTE: - CAN'T SCAN OR READ THE IMAGE

text_recogniton_chat-master.zip: - In this program we have to Input the Message and Chatbot will give the answer to the Followings Questions. 
NOTE : - I DON'T USED ANY API KEY'S